Fix -Wtype-limits and -Wsign-compare warnings
- WString::lastIndexOf: drop the `fromIndex < 0` guards; fromIndex is
  unsigned int so the comparison is always false (the `>= len` check
  already covers out-of-range). Behaviour unchanged.
- nonstd.h dtostrf: cast precision to int in the rounded-digit check to
  match the existing `(int)precision` style and silence the signed/
  unsigned comparison.

The default build was already warning-free; these only surfaced under
-Wall -Wextra. The remaining -Wextra warnings are intentional unused
parameters in the hardware no-op stubs.

+++ src/nonstd.h
@@ -169,7 +169,7 @@ char * dtostrf(float val, int width, unsigned int precision, char *buf)
 			s = fcvt(val, newPrecision, &newDecimalPoint, &sign);
 
 			// if rounded up to new digit (e.g. 0.09 to 0.1), move decimal point
-			if (newDecimalPoint - decpt == precision + 1) decpt++;
+			if (newDecimalPoint - decpt == (int)precision + 1) decpt++;
 		}
 	}
